feature density computations complete, ready for merge to main branch
